requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
catboost==1.2.8
category_encoders==2.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
choreographer==1.0.9
click==8.2.1
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.3
dash==3.2.0
debugpy==1.8.16
decorator==5.2.1
deprecation==2.1.0
executing==2.2.0
fastjsonschema==2.21.2
Flask==3.1.1
fonttools==4.59.1
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
idna==3.10
imbalanced-learn==0.14.0
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kaleido==1.0.0
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.44.0
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
narwhals==2.1.2
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.61.2
numpy==1.26.4
orjson==3.11.2
packaging==25.0
pandas==2.1.4
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
prompt_toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycaret==3.3.2
pydeck==0.9.1
Pygments==2.19.2
pyod==2.0.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.4
retrying==1.4.2
rpds-py==0.27.0
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
simplejson==3.20.1
six==1.17.0
sktime==0.26.0
smmap==5.0.2
stack-data==0.6.3
statsmodels==0.14.5
streamlit==1.48.1
tbats==1.1.3
tenacity==9.1.2
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wurlitzer==3.1.1
xgboost==3.0.4
xxhash==3.5.0
yellowbrick==1.5
zipp==3.23.0
